+++ jdbc/src/main/java/clazzfish/jdbc/JdbcStarter.java
@@ -62,12 +62,11 @@ public final class JdbcStarter {
      */
     public static void record() {
         start();
-        ConnectionMonitor.getInstance().addMeAsShutdownHook();
         SqlStatistic statistic = SqlStatistic.getInstance();
         statistic.setXPorter(XPorter.createCsvXPorter(dumpURI));
         statistic.registerMeAsMBean();
         statistic.addMeAsShutdownHook();
-        log.trace("ConnectionMonitor and {} are registered as shutdown hook.", statistic);
+        log.trace("{} is registered as shutdown hook.", statistic);
     }
 
     /**