+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+
+### Fixed
+
+- compact output for ConnectionMonitor at shutdown
+  ([issue #50](https://github.com/oboehm/ClazzFish/issues/50))
+
 ## [3.0] (01-Jan-2026)
 
 ### Changed